pyproject.toml
[tool.poetry]
name = "habittracker"
version = "0.1.0"
description = ""
authors = ["serebriakovanton <cu-hawk@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
celery = "^5.4.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
redis = "^5.0.8"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
django-filter = "^24.3"
setuptools = "^73.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
